refactor(channels): split QQ adapters
